Implemented BagResizingArray (ResizingArrayBag)
Added all the methods from the Java library. Included the Enumerable
library to make the class slightly more useable from a ruby perspective.

Added tests for the methods.

Added documentation for the methods.

Implemented the testing client to have the same behaviour as the
original Java library.

Added license information.