Implemented StackResizingArray (ResizingArrayStack)
Added all the methods of the original Java library. Included the Enumerable
library to make the class slightly more useable from a ruby perspective.

Added tests for all the methods.

Added documentation for all the methods.

Implemented the testing client to have the same behaviour as the
original Java library.

Added license information.