Implemented Stopwatch (Stopwatch)
Added all the methods from the original Java library.

Added tests for all the methods.

Added documentation for all the methods.

Added a testing client to approximate the Java behaviour. It's not
possible to duplicate the exact behaviour given the lack of a Math.pow
function. As an alternative, **0.5 was used.

Added license information.