Implemented Insertion (Insertion)
Implemented the various sort overloads. Where the Java versions took a
low and a high parameter for the array slices, the Ruby versions take a
Range. This was seen as both slightly more flexible, and a bit more
transparent to the user (their is no need to remember if the low and
high values are inclusive). Where the Java versions took a comparator,
the ruby versions take a blook as this seems like both the most Ruby
thing to do and the most flexible alternative to Java's comparators.

Implemented an additional non-mutating version of the sort which makes
a copy of the array before sorting. The sort returns the copy and leaves
the source array unmutated.

Added documentation for all the implemented methods.

Added tests for all the implemented methods.

Added a testing client that has the same behaviour as the original Java
library.

Added license information.